backend/requirements.txt
fastapi
pydantic
uvicorn
openai==0.27.6
python-dotenv
